+++ src/Container/VanillaContainer.php
@@ -10,7 +10,7 @@ namespace Rougin\Slytherin\Container;
  * @package Slytherin
  * @author  Rougin Royce Gutib <rougingutib@gmail.com>
  */
-class VanillaContainer implements \Interop\Container\ContainerInterface
+class VanillaContainer implements ContainerInterface
 {
     /**
      * @var array
@@ -19,15 +19,15 @@ class VanillaContainer implements \Interop\Container\ContainerInterface
 
     /**
      * Adds a new instance to the container.
+     * NOTE: To be removed in v1.0.0
      *
-     * @param string $id
-     * @param mixed  $concrete
+     * @param  string     $alias
+     * @param  mixed|null $concrete
+     * @return self
      */
-    public function add($id, $concrete)
+    public function add($alias, $concrete = null)
     {
-        $this->instances[$id] = $concrete;
-
-        return $this;
+        return $this->set($alias, $concrete);
     }
 
     /**
@@ -46,28 +46,42 @@ class VanillaContainer implements \Interop\Container\ContainerInterface
     /**
      * Finds an entry of the container by its identifier and returns it.
      *
-     * @param  string $id Identifier of the entry to look for.
+     * @param  string $alias Identifier of the entry to look for.
      * @return mixed
      */
-    public function get($id)
+    public function get($alias)
     {
-        if (! $this->has($id)) {
+        if (! $this->has($alias)) {
             $message = 'Alias (%s) is not being managed by the container';
 
-            throw new Exception\NotFoundException(sprintf($message, $id));
+            throw new Exception\NotFoundException(sprintf($message, $alias));
         }
 
-        return $this->instances[$id];
+        return $this->instances[$alias];
     }
 
     /**
      * Returns true if the container can return an entry for the given identifier.
      *
-     * @param  string $id
+     * @param  string $alias
      * @return boolean
      */
-    public function has($id)
+    public function has($alias)
     {
-        return isset($this->instances[$id]);
+        return isset($this->instances[$alias]);
+    }
+
+    /**
+     * Sets a new instance to the container.
+     *
+     * @param  string     $alias
+     * @param  mixed|null $concrete
+     * @return self
+     */
+    public function set($alias, $concrete = null)
+    {
+        $this->instances[$alias] = $concrete;
+
+        return $this;
     }
 }